Split node.rs into multiple files
